Move all predicates to a package and update import statements
