--- src/main/java/seedu/address/model/person/EmailContainsKeywordsPredicate.java
@@ -1,34 +1,0 @@
-package seedu.address.model.person;
-
-import seedu.address.commons.util.StringUtil;
-
-/**
- * Tests that a {@code Person}'s {@code Email} matches any of the keywords given.
- */
-public class EmailContainsKeywordsPredicate extends FieldPredicates {
-
-    public EmailContainsKeywordsPredicate(String keyword) {
-        super(keyword);
-    }
-
-    @Override
-    public boolean test(Person person) {
-        return keywords.stream()
-                .anyMatch(keyword -> StringUtil.containsWordIgnoreCase(person.getEmail().value, keyword));
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        if (other == this) {
-            return true;
-        }
-
-        // instanceof handles nulls
-        if (!(other instanceof EmailContainsKeywordsPredicate)) {
-            return false;
-        }
-
-        EmailContainsKeywordsPredicate otherNameContainsKeywordsPredicate = (EmailContainsKeywordsPredicate) other;
-        return keywords.equals(otherNameContainsKeywordsPredicate.keywords);
-    }
-}
