--- src/main/java/seedu/address/model/person/TagContainsKeywordsPredicate.java
@@ -1,34 +1,0 @@
-package seedu.address.model.person;
-
-import seedu.address.commons.util.StringUtil;
-
-/**
- * Tests that a {@code Person}'s {@code Tags} matches any of the keywords given.
- */
-public class TagContainsKeywordsPredicate extends FieldPredicates {
-    public TagContainsKeywordsPredicate(String keyword) {
-        super(keyword);
-    }
-
-    @Override
-    public boolean test(Person person) {
-        return keywords.stream()
-                .anyMatch(keyword -> person.getTags().stream().anyMatch(
-                        tag -> StringUtil.containsWordIgnoreCase(tag.tagName, keyword)));
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        if (other == this) {
-            return true;
-        }
-
-        // instanceof handles nulls
-        if (!(other instanceof TagContainsKeywordsPredicate)) {
-            return false;
-        }
-
-        TagContainsKeywordsPredicate otherNameContainsKeywordsPredicate = (TagContainsKeywordsPredicate) other;
-        return keywords.equals(otherNameContainsKeywordsPredicate.keywords);
-    }
-}
